Reorganize tests: dissolve CoverageGapSpec into proper homes
